In --no-daemon unlink of nonexisting pidfile
When running application using --no-deamon, when process is killed with signal SIGTERM, it exits with code 1, due to fact that pidfile does not exist

+++ loader.js
@@ -187,7 +187,9 @@ Loader.stop = function () {
 	killWorkers();
 
 	// Clean up the pidfile
-	fs.unlinkSync(pidFilePath);
+	if (nconf.get('daemon') !== 'false' && nconf.get('daemon') !== false) {
+		fs.unlinkSync(pidFilePath);
+	}
 };
 
 function killWorkers() {
